% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpois_continuous.R
\name{dpois3_continuous}
\alias{dpois3_continuous}
\title{continuous binomial distribution approximated by continuous Poisson distribution}
\usage{
dpois3_continuous(k, n, p)
}
\arguments{
\item{k}{Number of successes}

\item{n}{Number of trials}

\item{p}{Probability of success}
}
\value{
The probability of k successes in n trials with probability p
}
\description{
continuous binomial distribution approximated by continuous Poisson distribution
}
\examples{
  dpois3_continuous(3, 10, 0.3)
  dpois3_continuous(0, 300, 0.01)
}
