% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpois_continuous.R
\name{dpois_continuous}
\alias{dpois_continuous}
\title{Poisson distribution, version continuous in both arguments}
\usage{
dpois_continuous(x, lambda)
}
\arguments{
\item{x}{the number of events that actually occur in a given interval of
time or space; x >= 0}

\item{lambda}{the number of events that on average occur in that given
interval; lambda >= 0}
}
\value{
The probability of x events occurring in the interval or space given
that lambda events occur on average
}
\description{
This is a continuous version of dpois(x, lambda).
In dpois(x, lambda) x is restricted to be a non negative integer;
lambda can be any non negative real number.
In dpois_continuous(x, lambda) both x and lambda can be any non negative
real number.
}
\examples{
  dpois_continuous(3, 3)
}
