% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpois_continuous.R
\name{mpois_continuous}
\alias{mpois_continuous}
\title{Under construction: Bayesian upper, i.e. Maximum, credible bound for defect rate for Poisson model}
\usage{
mpois_continuous(k, n, certainty)
}
\arguments{
\item{k}{A numeric value representing the number of observed defects.
Need not be an integer. Must be non-negative.}

\item{n}{A numeric value representing the sample size. Need not be an integer.
Must be greater than 0.}

\item{certainty}{A numeric value between 0 and 1 representing the desired
level of confidence (e.g., 0.95 for 95\% certainty).}
}
\value{
A numeric value representing the maximum estimated defect rate (p).
}
\description{
This function calculates the maximum estimated defect rate (p) given
minimal certainty of \code{certainty}, for a population,
which was sampled with \code{n} samples of which
the sum of defects rates is \code{k}.
The calculation is based on the Poisson model.

In other words, the function finds the value of \code{p} that solves the equation:
\code{P(X <= k | n, p) = 1 - certainty}
where \code{X} follows a Poisson distribution.
}
\keyword{internal}
