% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpois_continuous.R
\name{dpois3_continuous}
\alias{dpois3_continuous}
\title{Poisson density function, version continuous in both arguments}
\usage{
dpois3_continuous(k, n, p)
}
\arguments{
\item{k}{Number of successes}

\item{n}{Number of trials}

\item{p}{Probability of success}
}
\value{
The probability of k successes in n trials with probability p
based on the Poisson model.
}
\description{
Poisson density function, version continuous in both arguments
}
\details{
dpois3(k, n, p)
= dpois_continuous(k, n * p)
= lambda^x * exp(-lambda) / gamma(x + 1)
Here x = k and lambda = n\emph{p.
So
dpois3(k, n, p)
= (n}p)^k * exp(-n*p) / gamma(k + 1)

We can compare this to dgamma(x, shape, rate).
Here shape = k + 1 and rate = n. And rate = 1/scale.

dgamma(x, shape, rate)
= ((scale^shape) * gamma(shape))^-1 * x^(shape-1) * exp(-(x/scale))
divide instead of ^-1
= x^(shape-1) * exp(-(x/scale)) / ((scale^shape) * gamma(shape))
fill in shape
= x^(k + 1 -1) * exp(-(x/scale)) / ((scale^(k + 1)) * gamma(k + 1))
k + 1 -1  = k
= x^k * exp(-(x/scale)) / ((scale^(k + 1)) * gamma(k + 1))
scale = 1/n
= x^k * exp(-(x\emph{n)) / (((1/n)^(k + 1)) * gamma(k + 1))
(1/n)^(k + 1) = (n^(k+1))^-1 = 1/(n^(k+1))
= x^k * exp(-(x}n)) * (n^(k + 1)) / gamma(k + 1)
x = p
= p^k * exp(-(p\emph{n)) * (n^(k + 1)) / gamma(k + 1)
exp(-(p}n)) = exp(-n\emph{p)
= p^k * exp(-n}p) * (n^(k + 1)) / gamma(k + 1)
remove some ()
= p^k * exp(-n\emph{p) * n^(k + 1) / gamma(k + 1)
p^k * n^(k + 1) = n * (n}p)^k
= n * (n\emph{p)^k * exp(-(n}p)) / gamma(k + 1)
}
\examples{
  dpois3_continuous(3, 10, 0.3)
  dpois3_continuous(0, 300, 0.01)
}
